Merged PR 3113: Refactor loading spinner
In this PR the Overlay got a refactor. First A ReactNode was stored in the global store and is changed to a custom Type.

The loading spinner got a dot animation

There is a new component that displays the loading spinner in full screen.

All the z-indexes are set from one file now