Merged PR 3140: Place bid logic + service + listener
- Backend logic for placing a bid
- Frontend request and listener for placing a bid
- Storing players' bids in match state on both backend and frontend
- Hook for accessing latest bid on frontend
- Showing loading overlay when awaiting response from server after placing a bid

When testing, you will have to input the values through temporary browser prompts that will pop up when clicking "Bid". In the browser console you will be able to read the latest bid after it has been processed. In case of invalid input the player turn should not advance to the next one and you should be able to re-place your bid.

Related work items: #13135