Merged PR 3167: Fix bug reload page
###QA

Refresh the page preferably a couple of times during the stages of a match. The expected behaviour is that the player will reconnect to the match and will still be able to perform the actions the respective stage. Furthermore, upon inspecting the page, in Application > Local Storage, the match variables should be deleted after the match is ended or when the player logs out.
Lastly, the player should be able to join/start a new game.

###Description
This PR solves the issue of a player being disconnected from the game when the page is reloaded at any given time. The game data is now being stored in the browser local storage during a match and is loaded into the store whenever the page is reloaded.
The matchId parameter is now stored in the lobby stage and is not reinitialized from the URL params during the match anymore.
playersLeft field added to the backend's state for keeping a record of players that disconnected from a match.

Note: The "Go For It" button can be retriggered by the user if the page is reloaded, although to no effect.

Related work items: #13760