Merged PR 3221: Fix bug on joining a second match
The local storage is now cleared if the matchId extracted from the URL doesn't match the current one (case when user joins a new match).

Note: Clearing the local storage when the Match Component is unmounted is not a suitable solution, as the component is unmounted when the page is refreshed as well.

QA
In the endOfMatchStage, the user should be able to enter a new lobby via a link.

Related work items: #14508