Merged PR 3240: Profanity moderation for account registration
Added plugin for moderating profanity. Can be integrated in the future for the chat system as well. bad-words is the most popular project, no shady dependencies and the plugin size is approx. 11.17 KB.

If the user tries to register a new account with an inappropriate username, the backend will return an INVALID_ARGUMENT Nakama Error that will be translated to the "Username contains profanity" message that will be displayed to the user.

###QA
If a user inputs a username that contains profanity, the error message "Username contains profanity" should be displayed, the account registration process should be interrupted.

Related work items: #14541