Merged PR 3355: Base Leo programs for Dice, PowerUp & Match
This PR introduces the game specific Leo programs that are responsible for creating (and burning) records for:
* Dice
* PowerUp
* Match Settings
* Match Summary

Due to limitations in Leo we can only have one record type per program, and each program can only have one "main" file in the source directory. So we end up with four different programs one for each record type.
In order to test you need to navigate in to the root of each of this programs and run `leo run {transitionName}`
If this is called with no input it will run with the test input present in the ".in" files in the "inputs" directory