Merged PR 3389: Timer implementation on the backend
The current status of this PR is an implimentation of the backend logic.

Timer service whicht keeps track of time based on the game loop ticks.
The amount of time per round is based the round stage

Known Issues:

Reloading on the result page is not working proparly

Related work items: #13144