Merged PR 3394: Updated views for Boloney and Exact actions
Updated views for boloney and exact actions. QA setup requires 3 players as the view for the idle player in case of the boloney action was also changed.

The sizes of the png files are at most 200 kb.

Disabled Exact & Boloney buttons for the first bidder.

Exact
https://www.figma.com/file/UB1htt9Flaql9UyxjsXnvb/Dice-game?node-id=651%3A10681&t=uegffGWiRZiSN1qJ-0

Boloney
https://www.figma.com/file/UB1htt9Flaql9UyxjsXnvb/Dice-game?node-id=1378%3A78204&t=uegffGWiRZiSN1qJ-0

Related work items: #15029, #15030, #15617