Merged PR 3497: Power-up service structure
- defined types for power-up payloads
- defined backend service structure
- added new op-code and notification code for power-ups
- implemented basic bird's eye logic (untested since i need the frontend for that)

This PR is mainly structural, so no need to QA. Certain behaviours need frontend improvements in order to be tested and properly developed. There are still a few temporary things on the frontend which are useful for testing but will be deleted after the final implementation, I've placed TODOs in front of all of them. Take your time to get familiar with this structure since we'll have to develop using that. Suggestions are accepted.

Related work items: #15723