Merged PR 3502: Heal dice action
This feuture wil introduce the game logic for healing a dice.

QA:
start a game with initial power-ups.
when its your turn press the heal-dice button
Mark the power-ups that you want to exchange for a new Die
press continue and you should get a new dice an lose the selected powerUps

For the other players the should receive a Notification and an update player side bar.

Related work items: #13131