Merged PR 3554: Power up list fix
Fixes the issues of displaying the power ups, in the modals and when healing dice.

Check the power up modal from the hud for how it displays the power ups, for 10 power ups and 3 power ups.

Check the power up modal from the power up button when u pick an action for how it displays the power ups, for 10 power ups and 3 power ups.

Check the power ups when u heal a die for how it displays the power ups, for 10 power ups and 3 power ups.

Related work items: #16023