Merged PR 3604: Corrected alignments in page
Added hover to dice for placing bid/grill.
Repositioning of notifications when the bottom button is not present.
Centered dice container.
On match summary, the call to action was moved at the bottom of the page
Refreshing page after session expired redirects the user to the login.

Related work items: #16117, #16119, #16121, #16122