Merged PR 3641: Match History
Match History feature

Added flag for disabling the zk toolkit. Currently it's set to true for reducing the time of the QA.

QA

The events should look according to the figma designs, and the events should be in order.

Note: At least a power-up with and without target need to be tested, besides calling exact and boloney.

https://www.figma.com/file/UB1htt9Flaql9UyxjsXnvb/Boloney!?node-id=1936%3A110746&t=x73J54EiKEqxmeYh-0