Merged PR 3670: Power-up Second chance implementation
Now you can use the powerup Second chance!

When using the powerup you need to select three dice that you want to reroll.

QA:
The player is not able to reroll more then 3 die.
The player can roll les than 3 dice
When no dice are selected the player is not able to press the button

Related work items: #12771