Merged PR 3837: Refractor colors and opacity
Updated the application with the colors and opacity of the design system, only colors and opacity within the design system are used.
Renamed the colours to be more concise.
Added a markdown file for file structure of the upcoming design system.

QA
Check the entire application, to see if the colours + opacity look the same as before. Check figma for references.

Related work items: #17057, #17191