Merged PR 4150: Legal Content
###QA
####!!! Rebuild yarn

Added react-markdown dependency as the pages are in an MD format. Using this dependency is the easiest way to handle the legal content as the original form is in a google doc format. Custom components can be added for replacing the html tags.

A component for tables should be implemented in the future
Added 3 new pages for the legal content

Buttons in the footer of the contact page are now functional.

Related work items: #18676