Merged PR 4195: Avoid error from expired refresh token
Reset the auth tokens if the refresh token expires in order to avoid errors related to authentication.

In the future we might want to consider one of the recommendations on nakama's forum, namely to update the refresh token manually when joining a match if it reached a certain threshold (valid for less than 30 mins for instance).

Related work items: #18750