Merged PR 4247: Match Settings molecule
Match Settings and ToolTip molecules.

Passing as props the `active` useState of the tooltip is not possible as the same instance would be referenced for all tooltips. If anyone would a suggestion for addressing this problem I would love to hear it.

###QA
In `/test` route press on Menu & Match Settings and check that the figma designs are respected.