Merged PR 4239: History Molecule
Added history molecule. Added a new file in `./test` that should be used in the future for the history organism.
Added BottomScroller molecule as well. Unfortunately passing  the ref or the callback func in useEffect as props would introduce new dependencies on the useEffect that affect the functionality.

###QA
Check in `/test` route that the history entries look according to the designs.

Related work items: #17195